# Conflicts:
#	infra/src/main/kotlin/fi/fta/geoviite/infra/tracklayout/LocationTrackDao.kt

+++ infra/src/main/kotlin/fi/fta/geoviite/infra/tracklayout/LocationTrackDao.kt
@@ -428,7 +428,7 @@ class LocationTrackDao(
                   select *
                     from layout.alignment
                     where location_track.alignment_id = alignment.id
-                      and (:min_length is null or alignment.length>=:min_length)
+                      and (:min_length::numeric is null or alignment.length>=:min_length)
                       and location_track.alignment_version = alignment.version
                       and postgis.st_intersects(postgis.st_makeenvelope(:x_min, :y_min, :x_max, :y_max, :layout_srid),
                                                 alignment.bounding_box)
